#39 project rename and namespace fix
