+++ src/ControlPanel/ControlPanel.WebApi.Client.Unit.Tests/ControlPanel.WebApi.Client.Unit.Tests.csproj.DotSettings
